backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
alembic==1.12.0
python-telegram-bot==20.6
openai==1.3.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
psutil==5.9.6